Moving ReadArgumentCheckNode node to its own package
